almost final .

+++ lib/data/models/product_model.dart
@@ -12,7 +12,7 @@ class ProductModel {
   String? discountPrice;
   String? image;
   int? stock;
-  double? star;
+  int? star;
   String? remark;
   int? categoryId;
   int? brandId;
